requirements.txt
paho-mqtt==1.6.1
pyqt5==5.15.9
pocketsphinx==5.0.4
google-cloud-speech==2.25.1
google-cloud-texttospeech==2.14.1
sounddevice==0.5.2
soundfile==0.12.1
numpy==2.2.1
scipy==1.14.1
pandas==2.2.3
matplotlib==3.8.4
pyqtgraph==0.13.3
icecream==2.1.3